added TestBulkSyncUsercases test, changed file name
